requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
celery==5.3.4
redis==5.0.1

# ML & Data Science
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1

# API & Scraping
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# Chatbot & NLP
langchain==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
channels==4.0.0
daphne==4.0.0

# Visualization
bokeh==3.3.2
